+++ client/src/components/StudyReview/StudyReview.tsx
@@ -1,14 +1,45 @@
 import React from 'react';
 import { CardIdsAndCorrectness, FormattedCard } from '../../../../types';
-import StudyReviewListItem from '../StudyReviewListItem/StudyReviewListItem';
 import {ButtonType, ColouredButton} from '../ColouredButton/ColouredButton'
 import Divider from '../Divider/Divider';
 import { updateQuizBoxes } from '../../utils/axios';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
+import { faCheckCircle, faTimesCircle } from '@fortawesome/free-solid-svg-icons'
 
-type StudyReviewProps = {
-  quizId: string
-  cardList: FormattedCard[]
-  onFinishReview: () => void
+type StudyReviewListItemProps = {
+  reviewItem: FormattedCard
+}
+
+const studyReviewListItemStyles = {
+  iconContainer: {
+    marginRight: '3%',
+    marginLeft: '1%'
+  },
+  reviewContainer: {
+    display: 'flex',
+    flexDirection: 'row' as const,
+    margin: 'auto',
+    alignItems: 'center'
+  },
+  textContainer: {
+  }
+}
+
+const StudyReviewListItem = ({reviewItem}: StudyReviewListItemProps) => {
+  return (
+    <div style={studyReviewListItemStyles.reviewContainer} data-testid="study-review-list-item-container">
+      <div style={studyReviewListItemStyles.iconContainer}>{(correct: boolean) => {
+        return correct ? 
+          <FontAwesomeIcon icon={faCheckCircle} color="green" size="3x" title="icon for correct answer"/> : 
+          <FontAwesomeIcon icon={faTimesCircle} color="red" size="3x" title="icon for incorrect answer"/>
+      }}</div>
+      <div style={studyReviewListItemStyles.textContainer}>
+        <p>Prompt: {reviewItem.prompt}</p>
+        <p>Your Answer: {reviewItem.givenAnswer}</p>
+        <p>Actual Answer: {reviewItem.answer}</p>
+      </div>
+    </div>
+  )
 }
 
 const loadReviewList = (cardList: FormattedCard[]) => {
@@ -43,6 +74,12 @@ const calculateCorrectAnswers = (cardList: FormattedCard[]) => {
 
 const calculateCorrectPercentage = (cardList: FormattedCard[]) => {
   return ((calculateCorrectAnswers(cardList) / cardList.length) * 100).toFixed(2);
+}
+
+type StudyReviewProps = {
+  quizId: string
+  cardList: FormattedCard[]
+  onFinishReview: () => void
 }
 
 const StudyReview = ({quizId, cardList, onFinishReview}: StudyReviewProps) => {
