--- client/src/components/StudyReviewListItem/__mocks__/StudyReviewListItem.tsx
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const StudyReviewListItem = () => {
-  return (
-    <div data-testid="study-review-list-item-container"></div>
-  )
-}
-
-export default StudyReviewListItem
