+++ client/src/screens/Study/Study.test.tsx
@@ -5,8 +5,6 @@ import userEvent from '@testing-library/user-event'
 import { dropAllTestCollections, createQuiz} from '../../../../testUtils/testUtils';
 import Study from './Study';
 
-jest.mock('../../components/StudyReviewListItem/StudyReviewListItem');
-
 describe('Study', () => {
   afterAll(async () => {
     await dropAllTestCollections();
